+++ changelog.md
@@ -2,10 +2,11 @@
 - added RandLoops module
 - added RandLoops8 module
 - added Turing mode on trigSeq and TrigSeq+
-- added 4ms-MetaModule compatibility
 - trigSeq trigSeq+ added Turing mode, changed RST knob behavior according to Turing mode.
 - stepSeq stepSeq+ fixed bug on reverse direction 
 - stepSeq stepSeq+ changed range display on step knobs
+- adder8 changed range display on volt/atn knobs
+- added 4ms-MetaModule compatibility
 - improved samples memory management on sampler modules
 
 ### 2.7.0 (2025-02-01)